updatertest.Dependencies cleanup
