+++ updater/docker/dependencies_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/thepwagner/action-update-go/updater"
+	"github.com/thepwagner/action-update-go/updater/docker"
 	"github.com/thepwagner/action-update-go/updatertest"
 )
 
@@ -18,5 +19,6 @@ func TestUpdater_Dependencies(t *testing.T) {
 			{Path: "alpine", Version: "3.11.0"},
 		},
 	}
-	updatertest.DependenciesFixtures(t, updaterFromFixture, cases)
+	updaterFactory := func(root string) updater.Updater { return docker.NewUpdater(root) }
+	updatertest.DependenciesFixtures(t, updaterFactory, cases)
 }
