--- updater/docker/updater_test.go
@@ -1,23 +1,0 @@
-package docker_test
-
-import (
-	"fmt"
-	"testing"
-
-	deepcopy "github.com/otiai10/copy"
-	"github.com/stretchr/testify/require"
-	"github.com/thepwagner/action-update-go/updater"
-	"github.com/thepwagner/action-update-go/updater/docker"
-)
-
-func updaterFromFixture(t *testing.T, fixture string) updater.Updater {
-	tempDir := tempDirFromFixture(t, fixture)
-	return docker.NewUpdater(tempDir)
-}
-
-func tempDirFromFixture(t *testing.T, fixture string) string {
-	tempDir := t.TempDir()
-	err := deepcopy.Copy(fmt.Sprintf("testdata/%s", fixture), tempDir)
-	require.NoError(t, err)
-	return tempDir
-}
