+++ updater/dockerurl/dependencies_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/thepwagner/action-update-go/updater"
+	"github.com/thepwagner/action-update-go/updater/dockerurl"
 	"github.com/thepwagner/action-update-go/updatertest"
 )
 
@@ -16,5 +17,7 @@ func TestUpdater_Dependencies(t *testing.T) {
 			{Path: "github.com/elixir-lang/elixir/releases", Version: "v1.10.3"},
 		},
 	}
-	updatertest.DependenciesFixtures(t, updaterFromFixture, cases)
+
+	updaterFactory := func(root string) updater.Updater { return dockerurl.NewUpdater(root) }
+	updatertest.DependenciesFixtures(t, updaterFactory, cases)
 }
