+++ updater/dockerurl/updater_test.go
@@ -2,12 +2,8 @@ package dockerurl_test
 
 import (
 	"fmt"
-	"testing"
 
-	deepcopy "github.com/otiai10/copy"
-	"github.com/stretchr/testify/require"
 	"github.com/thepwagner/action-update-go/updater"
-	"github.com/thepwagner/action-update-go/updater/dockerurl"
 )
 
 //go:generate mockery --outpkg dockerurl_test --output . --testonly --name repoClient --structname mockRepoClient --filename mockrepoclient_test.go
@@ -25,15 +21,3 @@ var (
 	dep     = updater.Dependency{Path: depPath, Version: previousVersion}
 	update  = updater.Update{Path: depPath, Previous: previousVersion, Next: nextVersion}
 )
-
-func updaterFromFixture(t *testing.T, fixture string) updater.Updater {
-	tempDir := tempDirFromFixture(t, fixture)
-	return dockerurl.NewUpdater(tempDir)
-}
-
-func tempDirFromFixture(t *testing.T, fixture string) string {
-	tempDir := t.TempDir()
-	err := deepcopy.Copy(fmt.Sprintf("testdata/%s", fixture), tempDir)
-	require.NoError(t, err)
-	return tempDir
-}
