+++ updater/gomod/updater_test.go
@@ -4,23 +4,15 @@ import (
 	"fmt"
 	"testing"
 
-	deepcopy "github.com/otiai10/copy"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 	"github.com/thepwagner/action-update-go/updater"
 	"github.com/thepwagner/action-update-go/updater/gomod"
+	"github.com/thepwagner/action-update-go/updatertest"
 )
 
 func updaterFromFixture(t *testing.T, fixture string, opts ...gomod.UpdaterOpt) *gomod.Updater {
-	tempDir := tempDirFromFixture(t, fixture)
+	tempDir := updatertest.TempDirFromFixture(t, fixture)
 	return gomod.NewUpdater(tempDir, opts...)
-}
-
-func tempDirFromFixture(t *testing.T, fixture string) string {
-	tempDir := t.TempDir()
-	err := deepcopy.Copy(fmt.Sprintf("testdata/%s", fixture), tempDir)
-	require.NoError(t, err)
-	return tempDir
 }
 
 func TestMajorPkg(t *testing.T) {
