+++ updatertest/dependencies.go
@@ -9,12 +9,14 @@ import (
 	"github.com/thepwagner/action-update-go/updater"
 )
 
-type updaterFactory func(t *testing.T, fixture string) updater.Updater
+type updaterFactory func(root string) updater.Updater
 
+// DependenciesFixtures verifies .Dependencies() on an Updater initialized from a fixture.
 func DependenciesFixtures(t *testing.T, factory updaterFactory, cases map[string][]updater.Dependency) {
 	for fixture, expected := range cases {
 		t.Run(fixture, func(t *testing.T) {
-			u := factory(t, fixture)
+			tempDir := TempDirFromFixture(t, fixture)
+			u := factory(tempDir)
 			deps, err := u.Dependencies(context.Background())
 			require.NoError(t, err)
 			assert.Equal(t, expected, deps)
